kernel/will-it-scale: Add support to download libhwloc
On certain Linux distributions libhwloc may not be available. This
blocks the execution of will-it-scale.
Add support to compile and install libhwloc library using the hwloc
source code.

Update the will-it-scale source code to use the installed
libhwloc library. Restrict this functionality to the distros
thats do not ship libhwloc library

Signed-off-by: Sachin Sant <sachinp@linux.ibm.com>